# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: URJOB API
  description: "URJOB 프로젝트의 전체 API 명세 (Backend + BFF)"
  version: "1.2.0"
servers:
  - url: http://localhost:8081  # BFF 호스트 (dev)
    description: BFF (dev)
  - url: http://localhost:8082  # Backend 호스트 (dev)
    description: Backend (dev)
...
paths:
...
  /api/bff/corpmem/mydata:
    get:
      summary: "기업회원 대시보드 데이터 조회 (BFF)"
      description: "BFF가 백엔드의 회사 프로필 정보를 가져와 반환합니다."
      responses:
        '200':
          description: "회사 정보 조회 성공"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyInfo"
        '404':
          description: "회사 정보를 찾을 수 없음"

  # --- Auth ---
...
  /api/v1/jobs:
    get:
      summary: 채용 공고 목록 조회
      description: >
        현재 활성화된 모든 채용 공고를 반환한다.
        (검색, 페이징, 국가/지역, 직무 유형 필터 지원)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: searchType
          in: query
          schema:
            type: string
            enum: [title, content, title_content, writer]
          description: 검색 유형
        - name: query
          in: query
          schema:
            type: string
          description: 검색어
        - name: country
          in: query
          schema:
            type: string
          description: 국가 필터
        - name: location
          in: query
          schema:
            type: string
          description: 지역 필터
        - name: jobType
          in: query
          schema:
            type: string
            enum: [FULL_TIME, PART_TIME, CONTRACT, INTERNSHIP, TEMPORARY]
          description: 직무 유형 필터
      responses:
        "200":
          description: 채용 공고 목록 반환
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobListItem'
    post:
      summary: 채용 공고 등록
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreate'
      responses:
        "201":
          description: 채용 공고 등록 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
  /api/v1/jobs/{jobId}:
    get:
      summary: 채용 공고 상세 조회
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 채용 공고 상세 정보 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDetailItem'
        "404":
          description: 채용 공고를 찾을 수 없음
    put:
      summary: 채용 공고 수정
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobUpdate'
      responses:
        "200":
          description: 채용 공고 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
    delete:
      summary: 채용 공고 삭제
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 채용 공고 삭제 성공
...
  /api/v1/jobs/{jobId}/status:
    put:
      summary: 채용 공고 상태 변경
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [OPEN, CLOSED, EXPIRED, DRAFT]
              required:
                - status
      responses:
        "200":
          description: 채용 공고 상태 변경 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
  
  # ---------------------------
  # Corpmem (기업회원 관련)
  # ---------------------------
  /api/v1/corpmem/accept:
    post:
      summary: 승인 요청 처리
      description: "기업 회원의 승인 요청을 처리하는 API"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                  example: 123
                status:
                  type: string
                  example: "approved"
      responses:
        "200":
          description: 승인 요청이 처리됨
        "400":
          description: 잘못된 요청
...
  /api/v1/corpmem/corpinfo:
    get:
      summary: 회사 프로필 정보 조회
      description: "DB에 시드된 회사 프로필 데이터를 반환합니다."
      responses:
        "200":
          description: 회사 정보 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyInfo"
        "404":
          description: 회사 정보를 찾을 수 없음
    put:
      summary: 기업 정보 수정
      description: "기업 회원의 정보를 수정하는 API"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInfo"
      responses:
        "200":
          description: 기업 정보 수정 완료
        "400":
          description: 잘못된 요청
  /api/v1/corpmem/talents/filter:
    post:
      summary: 핵심역량 필터 기반 인재 검색
      description: "선택된 핵심역량을 바탕으로 검색된 인재 목록 반환"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                selectedSkills:
                  type: array
                  items:
                    type: string
                  example: ["계획성", "성실성"]
      responses:
        "200":
          description: 필터링된 인재 목록 반환
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TalentItem'

  # ---------------------------
  # Resumes (이력서)
  # ---------------------------
  /resumes:
    get:
      summary: 이력서 목록 조회
      responses:
        "200":
          description: 이력서 목록 반환
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResumeListItem'
    post:
      summary: 이력서 작성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResumeDetail'
      responses:
        "201":
          description: 이력서 작성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeDetail'
  /resumes/{id}:
    get:
      summary: 이력서 상세 조회
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: 이력서 상세 정보 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeDetail'
    put:
      summary: 이력서 수정
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResumeDetail'
      responses:
        "200":
          description: 이력서 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeDetail'
    delete:
      summary: 이력서 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: 이력서 삭제 성공

  # ---------------------------
  # Introductions (자기소개서)
  # ---------------------------
  /introductions:
    get:
      summary: 자기소개서 목록 조회
      responses:
        "200":
          description: 자기소개서 목록 반환
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Introduction'
    post:
      summary: 자기소개서 작성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntroductionCreate'
      responses:
        "201":
          description: 자기소개서 작성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Introduction'
  /introductions/{id}:
    get:
      summary: 자기소개서 상세 조회
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: 자기소개서 상세 정보 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Introduction'
    put:
      summary: 자기소개서 수정
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntroductionCreate'
      responses:
        "200":
          description: 자기소개서 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Introduction'
    delete:
      summary: 자기소개서 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: 자기소개서 삭제 성공

  # ---------------------------
  # Community (사용자 커뮤니티)
  # ---------------------------
  /community/posts:
    get:
      summary: 커뮤니티 게시글 목록 조회
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: country
          in: query
          schema:
            type: string
          description: (선택) 게시글 필터링용 나라 정보
        - name: searchType
          in: query
          schema:
            type: string
            enum: [title, content, title_content, writer]
          description: 검색 유형
        - name: query
          in: query
          schema:
            type: string
          description: 검색어
      responses:
        "200":
          description: 게시글 목록 반환
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  posts:
                    type: array
                    items:
                      $ref: '#/components/schemas/CommunityPostListItem'
    post:
      summary: 게시글 작성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommunityPostCreate'
      responses:
        "201":
          description: 게시글 작성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityPostDetail'
  /community/posts/{id}:
    get:
      summary: 게시글 상세 조회 (조회수 증가)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: 게시글 상세 정보 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityPostDetail'
    put:
      summary: 게시글 수정
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommunityPostCreate'
      responses:
        "200":
          description: 게시글 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityPostDetail'
    delete:
      summary: 게시글 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: 게시글 삭제 성공
  /community/posts/{postId}/comments:
    post:
      summary: 댓글 작성
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommunityCommentCreate'
      responses:
        "201":
          description: 댓글 작성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityComment'
  /community/posts/{postId}/comments/report:
    post:
      summary: 댓글 신고
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
              required:
                - content
      responses:
        "200":
          description: 댓글 신고 성공
  /community/posts/{postId}/comments/block:
    post:
      summary: 댓글 차단
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                commentId:
                  type: number
              required:
                - commentId
      responses:
        "200":
          description: 댓글 차단 성공

  # ---------------------------
  # JobOffers (구인자의 포지션 제안)
  # ---------------------------
...
components:
  schemas:
...
    JobListItem:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        company:
          type: string
        location:
          type: string
        salary:
          type: number
      required:
        - id
        - title
        - company
        - location
        - salary
    JobDetailItem:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        description:
          type: string
        requirements:
          type: string
        location:
          type: string
        company:
          type: string
        salary:
          type: number
        views:
          type: integer
        likes:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - requirements
        - location
        - company
        - salary
    Job:
      type: object
      properties:
        id:
          type: number
        employerId:
          type: number
        title:
          type: string
        description:
          type: string
        requirements:
          type: string
        location:
          type: string
        countryCode:
          type: string
        jobType:
          type: string
        salary:
          type: number
        views:
          type: integer
        likes:
          type: integer
        status:
          type: string
          enum: [OPEN, CLOSED, EXPIRED, DRAFT]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - employerId
        - title
        - description
        - requirements
        - location
        - jobType
        - salary
        - status
    JobCreate:
      type: object
      properties:
        employerId:
          type: number
        title:
          type: string
        description:
          type: string
        requirements:
          type: string
        location:
          type: string
        countryCode:
          type: string
        jobType:
          type: string
          enum: [FULL_TIME, PART_TIME, CONTRACT, INTERNSHIP, TEMPORARY]
        salary:
          type: number
      required:
        - employerId
        - title
        - description
        - requirements
        - location
        - jobType
        - salary
    JobUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        requirements:
          type: string
        location:
          type: string
        countryCode:
          type: string
        jobType:
          type: string
          enum: [FULL_TIME, PART_TIME, CONTRACT, INTERNSHIP, TEMPORARY]
        salary:
          type: number
        status:
          type: string
          enum: [OPEN, CLOSED, EXPIRED, DRAFT]
      required:
        - title
        - description
        - requirements
        - location
        - jobType
        - salary
    CompanyInfo:
      type: object
      properties:
        companyName:
          type: string
          example: "(주)유어잡"
        employeeCount:
          type: integer
          example: 100
        capitalAmount:
          type: number
          format: double
          example: 288
        revenueAmount:
          type: number
          format: double
          example: 5000
        netIncome:
          type: number
          format: double
          example: 500
        address:
          type: string
          example: "서울특별시 강남구"
        website:
          type: string
          nullable: true
          example: "https://www.yourjob.com"
        logo_url:
          type: string
          nullable: true
          example: "/img/logo.png"
      required:
        - companyName
        - employeeCount
        - capitalAmount
        - revenueAmount
        - netIncome
        - address
...
    ResumeListItem:
      type: object
      properties:
        id:
          type: number
          description: resumes.resume_id
        title:
          type: string
          description: resumes.resume_title
        date:
          type: string
          format: date-time
          description: resumes.created_at
      required:
        - id
        - title
        - date
    ResumeDetail:
      type: object
      properties:
        id:
          type: number
          description: resumes.resume_id
        jobSeekerId:
          type: number
          description: resumes.job_seeker_id
        title:
          type: string
          description: resumes.resume_title
        filePath:
          type: string
          description: resumes.file_path
        createdAt:
          type: string
          format: date-time
          description: resumes.created_at
        updatedAt:
          type: string
          format: date-time
          description: resumes.updated_at
      required:
        - id
        - jobSeekerId
        - title
        - filePath
        - createdAt
        - updatedAt
    Introduction:
      type: object
      properties:
        id:
          type: number
          description: 자기소개서의 고유 ID
        jobSeekerId:
          type: number
          description: 구직자의 ID
        title:
          type: string
          description: 자기소개서 제목
        content:
          type: string
          description: 자기소개서 내용
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - jobSeekerId
        - title
        - content
        - createdAt
        - updatedAt
    IntroductionCreate:
      type: object
      properties:
        jobSeekerId:
          type: number
        title:
          type: string
        content:
          type: string
      required:
        - jobSeekerId
        - title
        - content
    CommunityPostListItem:
      type: object
      properties:
        id:
          type: number
          description: community_posts.post_id
        subTitle:
          type: string
          description: "게시글 내용의 요약 (프론트엔드에서 계산)"
        title:
          type: string
          description: community_posts.title
        writer:
          type: string
          description: "게시글 작성자 이름"
        date:
          type: string
          format: date-time
          description: community_posts.created_at
        views:
          type: integer
          description: community_posts.views
        likes:
          type: integer
          description: community_posts.likes
        isMine:
          type: boolean
          description: "현재 사용자와 작성자 비교 결과"
        isNotice:
          type: boolean
          description: "공지 여부"
      required:
        - id
        - title
        - writer
        - date
        - views
        - likes
        - isMine
    CommunityPostDetail:
      type: object
      properties:
        id:
          type: number
          description: community_posts.post_id
        subTitle:
          type: string
          description: "게시글 내용 요약 또는 부제목"
        title:
          type: string
        writer:
          type: string
        date:
          type: string
          format: date-time
        views:
          type: integer
        likes:
          type: integer
        content:
          type: string
          description: community_posts.content
        comments:
          type: array
          items:
            $ref: '#/components/schemas/CommunityComment'
      required:
        - id
        - title
        - writer
        - date
        - views
        - likes
        - content
        - comments
    CommunityPostCreate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        area:
          type: string
      required:
        - title
        - content
        - area
    CommunityComment:
      type: object
      properties:
        id:
          type: number
          description: community_comments.comment_id
        writer:
          type: string
          description: "댓글 작성자 이름"
        date:
          type: string
          format: date-time
          description: community_comments.created_at
        content:
          type: string
          description: community_comments.content
        likes:
          type: integer
          description: community_comments.likes
        recommentId:
          type: number
          description: "부모 댓글 ID (대댓글인 경우)"
      required:
        - id
        - writer
        - date
        - content
        - likes
    CommunityCommentCreate:
      type: object
      properties:
        content:
          type: string
      required:
        - content
...